LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-container-baseline-001-expected.txt

 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 
 

PASS .wrapper 1 
PASS .wrapper 2 
PASS .wrapper 3 
FAIL .wrapper 4 assert_equals: 
<div class="wrapper" style="writing-mode: vertical-lr;">
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
  <div class="grid">
    <div class="i1"></div>
    <div class="i2" data-offset-x="0"></div>
    <div class="i3"></div>
  </div>
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
</div>
offsetLeft expected 65 but got 75
FAIL .wrapper 5 assert_equals: 
<div class="wrapper" style="writing-mode: vertical-lr;">
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
  <div class="grid">
    <div class="i1"></div>
    <div class="i2" style="writing-mode: horizontal-tb;" data-offset-x="0"></div>
    <div class="i3"></div>
  </div>
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
</div>
offsetLeft expected 65 but got 75
PASS .wrapper 6 
PASS .wrapper 7 
FAIL .wrapper 8 assert_equals: 
<div class="wrapper" style="writing-mode: vertical-rl;">
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
  <div class="grid">
    <div class="i1"></div>
    <div class="i2" data-offset-x="75"></div>
    <div class="i3"></div>
  </div>
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
</div>
offsetLeft expected 65 but got 75
FAIL .wrapper 9 assert_equals: 
<div class="wrapper" style="writing-mode: vertical-rl;">
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
  <div class="grid">
    <div class="i1"></div>
    <div class="i2" style="writing-mode: horizontal-tb;" data-offset-x="75"></div>
    <div class="i3"></div>
  </div>
  <div style="display: inline-block; width: 20px; height: 10px; background: black" data-offset-x="65"></div>
</div>
offsetLeft expected 65 but got 75
PASS .wrapper 10 
PASS .wrapper 11 
